Refactor: 중복 css 제거
