+++ src/styles/pages.module.css
@@ -1,3 +1,10 @@
+.commonPage {
+  display: flex;
+  flex-direction: column;
+  width: 100%;
+  padding: 0 2.7rem;
+}
+
 .homePage {
   display: flex;
   flex-direction: column;
