+++ src/styles/projects/projectscontent.module.css
@@ -5,6 +5,7 @@
 
 .project {
   flex: 2;
+  margin-top: 3rem;
 }
 
 .projectImg {
