--- src/styles/projects/projectspage.module.css
@@ -1,6 +1,0 @@
-.projectsPage {
-  display: flex;
-  flex-direction: column;
-  width: 100%;
-  padding: 0 2.7rem;
-}
